изменение загрузки данных пользователя и сохраненных фильмов
